+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ## 0.0.5
 
 - Implement internal caches clearing
+- Implement a formatter
 
 ## 0.0.4